move the main loop inside the ouput module
The module manages a large part of the system. The UPnP system has its
own thread management without any link with glib.
Only the gstreamer threads need to use glib API and output already
managed the main loop.
This patch create a new API inside the output module (loop).
