+++ src/output_module.h
@@ -34,6 +34,7 @@ struct output_module {
 
 	// Commands.
 	int (*init)(void);
+	int (*loop)(void);
 	void (*set_uri)(const char *uri, output_update_meta_cb_t meta_info);
 	void (*set_next_uri)(const char *uri);
 	int (*play)(output_transition_cb_t transition_callback);
